delete dynamic_resource and split infer.h
